Booking process up to the payment page
