+++ app/controllers/MongoDBController.scala
@@ -1,10 +1,12 @@
 package controllers
 
-import java.util.Date
+import java.sql.Timestamp
+import java.util.{Calendar, Date}
 import javax.inject.Inject
 
 import play.api.cache.Cache
 import play.api.Play.current
+
 import scala.util.Try
 import scala.concurrent.{Await, Future}
 import play.api.mvc._
@@ -312,5 +314,24 @@ class MongoDBController @Inject()(val messagesApi: MessagesApi)(val reactiveMong
       }
   }
 
+
+  //before payment
+  def payment(name:String,email:String) = Action {implicit request=>
+    val total=request.session.get("total").get.toInt
+    val adult=request.session.get("adult").get.toInt
+    val child=request.session.get("child").get.toInt
+    val concession=request.session.get("concession").get.toInt
+    val seats=request.session.get("seats").get
+    val newseats=seats.split(',').toList.map(elem=>elem.split(' ').toList)
+    val moviename=request.session.get("moviename").get
+    val date=request.session.get("date").get
+    val time=request.session.get("time").get
+    val room=request.session.get("room").get
+    val currentTimestamp = new Timestamp(Calendar.getInstance.getTime.getTime).toString
+
+    val reservation=new Reservation(currentTimestamp,name,email,adult,child,concession,newseats,total,moviename,date,time,room,false)
+    //the price in there that you want the checkout button to have
+    Ok(views.html.payment(total.toString,reservation)).withSession(request.session+("name"->name)+("email"->email)+("reservationId"->currentTimestamp))
+  }
 
 }
