+++ conf/routes
@@ -49,7 +49,7 @@ GET        /userinfo                    controllers.MongoDBController.userInfo(s
 # Map static resources from the /public folder to the /assets URL path
 GET         /assets/*file               controllers.Assets.at(path="/public", file)
 
-GET         /payment                    controllers.Application.payment(name:String,email:String)
+GET         /payment                    controllers.MongoDBController.payment(name:String,email:String)
 
 # Sessions
 GET         /session                    controllers.Application.session()
